Refactor SessionManager into interface with InMemorySessionManager default impl
Extract SessionManager as an interface (keeping SESSION_COOKIE_NAME and
DEFAULT_SESSION_TTL_SECONDS constants + the public API) and rename the
in-memory implementation to InMemorySessionManager. All injection points
(AuthenticationFilter, SessionAuthenticationFilter, PinotQueryResource,
PinotSessionLoginResource) already reference the SessionManager type so they
are unaffected. BaseControllerStarter now instantiates InMemorySessionManager.
SessionManagerTest updated to construct InMemorySessionManager directly while
keeping variable types as SessionManager.

Addresses @shounakmk219 review feedback on PR #18933.
